+++ src/App.js
@@ -1,14 +1,14 @@
 import logo from './logo.svg';
 import './App.css';
 import {Fragment} from "react";
-//import EntryPage from './Pages/Login'
+import Login from './Pages/Login/Login'
 import { useEffect } from 'react';
 //import axios from "axios";
 
 function App() {
   return (
     <Fragment>
-      <App/>
+      <Login/>
     </Fragment>
   );
 }
