+++ src/Pages/Login/Login.css
@@ -1,0 +1,157 @@
+@import url('https://fonts.googleapis.com/css?family=Jura:400');
+@import url('https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap');
+* { font-family: Jura, Arial; font-weight: 400; }
+body { margin: 0; padding: 0; }
+h1,h2,h3,h4,h5,h6,p { margin: 0; }
+
+
+
+section#entry-page {
+    display: flex;
+    flex-direction: column;
+    grid-template-columns: 1fr minmax(200px,400px) 1fr;
+    grid-template-rows: 1fr minmax(auto,1fr) 1fr;
+    grid-row: auto;
+    grid-gap: 0px;
+    width: 100%;
+    height: 100vh;
+    background: linear-gradient(-45deg,#0a8a8a 0%,#380036 74%);
+    background-size: 400% 400%;
+    animation: Gradient 15s ease infinite;
+    box-sizing: border-box;
+    /*border-bottom: 8px;*/
+    align-items: center;}
+
+h7 {
+    font-family: 'Press Start 2P', cursive;
+    font-size: 70px;
+    text-align: center;
+    font-weight: 400;
+    border-width: 1px;
+    align-content: center;
+    margin-top: 90px;
+    /*margin-left: 400px; /*need help aligning*/
+
+}
+
+    form {
+        grid-column: 2;
+        grid-row: 2;
+        display: inline-block;
+        grid-gap: 10px;
+        margin: auto 0;
+        padding: 20px;
+        background-color: rgb(120,120,120);
+        border-radius: 10px;
+        box-shadow: 0 32px 64px rgba(0,0,0,0.2);
+        align-items: center;
+        margin-top: 80px;
+
+    }
+        h2 {
+            margin-bottom: 5px;
+            text-align: center;
+            text-shadow: 0 4px 16px 	#808080;
+            font-size: 30px;
+            font-weight: 100;
+        }
+        fieldset {
+            margin: 0;
+            background-color: #A0A0A0;
+            border: none;
+            border-radius: 5px;
+            box-shadow: 0 1px 3px rgba(0,0,0,0.2);}
+        legend {
+            padding: 5px;
+            background-color: #A0A0A0;
+            border-radius: 5px;
+        }
+        ul {
+            margin: 0;
+            padding: 0;}
+        li {
+            display: grid;
+            align-items: center;
+            margin: 10px;}
+        a {
+            color: #02c;
+        }
+        em {
+            grid-column: span 2;
+            text-align: center;
+            padding: 5px;
+        }
+        label {
+            text-align: left;
+            padding-bottom: 2px;
+        }
+        input {
+            padding: 5px;
+            border: 1px solid #989898;
+            background: 	#C8C8C8;
+            border-radius: 5px;}
+        input:hover {
+             border: 1px solid #aaf;
+         }
+        button {
+            padding: 10px;
+            border: 1px solid rgba(0, 0, 0, 0);
+            border-radius: 5px;
+            background: #6190ab;
+            box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+            margin-top: 10px;
+            margin-left: 45px;
+        }
+.b1 {
+    padding: 10px;
+    border: 1px solid rgba(0, 0, 0, 0);
+    border-radius: 5px;
+    background: #6bab63;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+    margin-right: 20px;
+}
+
+button:hover {
+    /*background-color: #eef;*/
+    border: 1px solid #aaf;
+}
+
+
+{}
+{}
+{}
+@media only screen and (min-width: 420px) {
+    form {}
+        h2 {
+            font-size: 40px;
+        }
+        fieldset {}
+            ul {}
+                li {
+                    grid-template-columns: 100px 1fr;}
+                    label {
+                        padding-right: 10px;
+                        padding-bottom: 0;
+                        text-align: right !important;
+                        }
+                }
+
+
+@keyframes Gradient {
+    0% {
+        background-position: 0% 50%
+    }
+    50% {
+        background-position: 100% 50%
+    }
+    100% {
+        background-position: 0% 50%
+    }
+}
+
+
+/*rgba(255,255,255,0.9)*/
+/*background-color: #7f5a83;*/
+/*background-image: linear-gradient(315deg, #7f5a83 0%, #0d324d 74%);*/
+
+/*-45deg,#0cbaba 0%, #380036 74%*/
