+++ src/index.css
@@ -10,4 +10,5 @@ body {
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
+  font-family: 'Press Start 2P', cursive;
 }
